seperated ProtestMap, ProtestInfo
