+++ src/components/ProtestInfo/ProtestInfo.style.js
@@ -1,0 +1,25 @@
+import styled from 'styled-components/macro';
+
+export const Info = styled.div`
+  position: relative;
+  padding: 20px 34px;
+  background: #ffffff;
+  box-shadow: 0px 4px 10px -1px rgba(0, 0, 0, 0.15);
+  border-radius: 4px;
+  margin-top: -30px;
+  z-index: 5;
+
+  @media (min-width: 600px) {
+    display: grid;
+    grid-template-columns: 1fr 200px;
+  }
+`;
+
+export const Title = styled.h1`
+  font-weight: bold;
+  font-size: 28px;
+  line-height: 47px;
+  color: #000000;
+  margin-bottom: 8px;
+`;
+export const Details = styled.div``;
