+++ src/components/ProtestInfo/index.js
@@ -1,0 +1,3 @@
+import ProtestInfo from './ProtestInfo';
+
+export default ProtestInfo;
