+++ src/components/ProtestMap/ProtestMap.style.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components/macro';
+import { Map } from 'react-leaflet';
+
+export const MapWrapper = styled(Map)`
+  width: 100%;
+  height: 256px;
+  z-index: 0;
+`;
