+++ src/components/ProtestMap/index.js
@@ -1,0 +1,3 @@
+import ProtestMap from './ProtestMap';
+
+export default ProtestMap;
